pyproject.toml
[project]
name = "startup-research-agent"
version = "1.0.0"
description = "Multi-agent AI system for researching startups in India"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Startup Research Team", email = "team@example.com"}
]

dependencies = [
    "crewai==1.3.0",
    "crewai-tools>=0.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "langchain-openai>=0.1.0",
    "langchain>=0.1.0",
    "tabulate>=0.9.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"



